Resolve review issues
