+++ workflows/pipe-common/test/test_available_instance_provider.py
@@ -13,11 +13,11 @@
 # limitations under the License.
 
 import logging
-
 from datetime import datetime
+
 from mock import MagicMock, Mock
 
-from pipeline.hpc.instance import Instance, AvailableInstanceProvider
+from pipeline.hpc.instance.provider import Instance, AvailableInstanceProvider
 
 logging.basicConfig(level=logging.DEBUG, format='%(asctime)s [%(threadName)s] [%(levelname)s] %(message)s')
