+++ workflows/pipe-common/test/test_scale_up_handler.py
@@ -18,7 +18,7 @@ from mock import MagicMock, Mock
 
 from pipeline.hpc.autoscaler import GridEngineScaleUpHandler
 from pipeline.hpc.host import MemoryHostStorage
-from pipeline.hpc.instance import Instance
+from pipeline.hpc.instance.provider import Instance
 from utils import assert_first_argument_contained
 
 try:
